Track Messages Written to SCE Sign
Resolves #228
